Fix statistics update in ANALYZE table for Hive transactional tables
This commit fixes update of statistics resulting from ANALYZE run on
Hive transactional tables.  This fix adds a new ANALYZE transaction
type, and adds an AcidTransaction object to the metastore APIs that
update/set table and partition statistics, calling into formerly unused
metastore APIs to update statistics for transactional tables.

This commit is drawn narrowly, in that it only uses the new APIs
for the case of ANALYZE on transactional tables.  As of today,
INSERT, DELETE and UPDATE do not update column statistics when
run on Hive transactional tables.  We could enable that behavior
with a small follow-on PR now that the necessary Hive metastore
APIs are accessible.