Add runner span under query that aggerates all stages
The Trino spans below the the Query span are Dispatch, Analyzer, Planner
and then a sequence of Stage spans. Placing the set of Stage spans under
a new Runner span allows all execution spans to be grouped together. In
addition exposing the Runner span to the ExchangeContext allows for the
possiblity of creating Exchange spans in the future to track exchange
specific resources.