Decrease logging verbocity
Getting exception in is a non-rare faililure mode.
E.g. we are talking to SPI provided Exchange implementation here, and
it is legal for those methods to throw excption due to infrastuctural
problems or because we noticed that further task execution makes no
sense as it alrady consumed data from upstream task which failed
mid-way.
We should not log at error level.
Also excpetion will not be lost as it is propagated to split runner and
eventually will surfase as task failure and also possibly query failure.